backend/preprocess/requirements.txt
Pillow>=10.2.0
numpy>=1.26.0
transformers>=4.37.0
torch>=2.2.0
torchvision>=0.17.0
einops>=0.7.0
shapely>=2.0.0
timm>=0.9.12
sympy>=1.12
tqdm>=4.66.0
diffusers==0.32.2
huggingface-hub>=0.19.0
accelerate>=1.5.2